Added jsp pages

+++ src/main/java/cs544/carrental/dao/PaymentDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import cs544.carrental.domain.Payment;
 
 public interface PaymentDao extends GenericDao<Payment> {
-	List<Payment> findByPaymentId(int paymentId);
+	List<Payment> findByPaymentId(long paymentId);
 //	Page<Payment> findAll(Pageable pageable);
 	List<Payment> findAll();
